refactor(rinkaku): split main.rs into responsibility-per-module structure
Break the 4341-line `rinkaku/src/main.rs` (production 1965 + tests 2376)
into a slim composition-root + responsibility-per-module structure. Every
extracted function keeps its body byte-for-byte; only its module location
changes. Tests move with the function they pin, sitting in the same file
as an inline `#[cfg(test)] mod tests { ... }` block.

## New layout

    rinkaku/src/
    ├── main.rs (274 lines)      composition root: `fn main()` only
    ├── cli.rs                    Cli / Command / Format definitions
    ├── display.rs                DisplayMode + resolve_display_mode
    ├── notes.rs                  apply_entry_pivot, entry_pivot_empty_note,
    │                             garbage_input_note, repo_outline_empty_note
    ├── pipeline.rs               run_base_pipeline, build_resolver,
    │                             changed_paths, resolve_generated_paths,
    │                             read_stdin_diff
    ├── generated_paths.rs        `.gitattributes`-driven generated-path
    │                             detection (ADR 0010)
    ├── test_util.rs              shared #[cfg(test)] helpers: run_git,
    │                             init_repo_with_committed_file
    ├── git/                      local git subprocess wrappers
    │   ├── commands.rs           run_git_diff, list_git_files,
    │   │                         list_repo_files_for_outline, resolve_repo_root
    │   ├── file_read.rs          read_working_tree_file, read_git_show_file
    │   └── cat_file_batch.rs     read_git_show_files_batch + 3 helpers
    └── github/                   GitHub-specific glue (PR + gh + remotes)
        ├── pr_arg.rs             PrArg + parse_pr_arg + parse_positive_pr_number
        ├── pr_info.rs            PrInfo + fetch_pr_info + parse_pr_view_json
        ├── remote.rs             parse_github_remote, github_remote_matches,
        │                         git_remote_origin_url
        ├── workdir.rs            resolve_pr_workdir + ghq/cache probe cascade
        └── base_sha.rs           resolve_pr_base_sha + fetch/exists helpers

## Motivation

- `main.rs` was well past ADR 0028's watch threshold; the reviewer could
  no longer hold its outline in one context window. Every touched
  submodule now sits at ≤ 575 lines.
- The `github/` grouping anticipates upcoming PR review-comment work
  (`gh pr comment`, review-thread posting): those additions can drop
  into `github/review.rs` / `github/comment.rs` at the same granularity
  as the existing PR-fetch code, rather than piling further onto a
  single monolithic file.

## Verification

- `make test` — 167 (rinkaku) + 351 (rinkaku_core) + 541 (rinkaku_tui)
  = 1059 tests pass; matches baseline exactly (no test dropped).
- `make lint` — `cargo fmt --all --check` and
  `cargo clippy --all-targets --all-features -- -D warnings` both clean.
- The single test-code change is `assert_eq!(false, actual.exclude_tests)`
  → `assert!(!actual.exclude_tests)` in `cli.rs`, since clippy's
  `bool_assert_comparison` lint (now `-D warnings`) rejects the original;
  no behavioral test change.
